app/requirements.txt
bcrypt
dedust
dataclasses_json
stonfi
email-validator
fastapi[all]
python-jose
passlib
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
python-multipart
SQLAlchemy
uvicorn[standart]
alembic
asyncpg
filetype
phonenumbers
pytoncenter
dexscreener
